fixed underline on NavBar and removed components
