+++ react/components/ActiveStatusBar.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function ActiveStatusBar () {
+  return <span className="progress-bar progress-bar--active">| </span>;
+}
+
+export default ActiveStatusBar;
