+++ react/components/InactiveStatusBar.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function InactiveStatusBar () {
+  return <span className="progress-bar progress-bar--inactive">| </span>;
+}
+
+export default InactiveStatusBar();
